requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
pyyaml
matplotlib
plotly
scipy
xgboost
